rewrite: Clean up ElementDefinitionNavigator static utility functions, move declarations into separate partial class file
