+++ src/Hl7.Fhir.Specification/Hl7.Fhir.Specification.Net40.csproj
@@ -83,6 +83,7 @@
     <Compile Include="..\Hl7.Fhir.Core\Support\StringExtensions.cs">
       <Link>Support\StringExtensions.cs</Link>
     </Compile>
+    <Compile Include="Specification\Navigation\ElementDefinitionNavigationFunctions.cs" />
     <Compile Include="Specification\Navigation\ElementNavigatorSlicingExtensions.cs" />
     <Compile Include="Specification\Terminology\ExpansionSizeTooBig.cs" />
     <Compile Include="Specification\Terminology\ITerminologyService.cs" />
